+++ src/imio/smartweb/core/webcomponents/src/components/Events/Events.jsx
@@ -62,7 +62,12 @@ function EventsView(props) {
 
     // set batch
     const callback = () => {
-        setBatchSize(batchSize + 5);
+        setFilters(prevFilters => {
+            return { 
+              ...prevFilters, 
+              b_size: batchSize + 5
+            }
+          })
     };
 
     // coditional list render